+++ shramko.maksim/P3/inputMtrx.hpp
@@ -5,7 +5,7 @@
 
 namespace shramko
 {
-  std::istream & inputMtrx(std::istream & in, int* mtrx, size_t cnt_row, size_t cnt_col, int& read);
+  std::istream & inputMtrx(std::istream & in, int* mtrx, size_t cnt_row, size_t cnt_col, size_t& read);
 }
 
 #endif
