+++ shramko.maksim/P3/main.cpp
@@ -51,7 +51,7 @@ int main(int argc, char** argv)
   if (argv[1][0] == 1)
   {
     int mtrx[10000] = {0};
-    int read = 0;
+    size_t read = 0;
     if (!shramko::inputMtrx(finput, mtrx, cnt_row, cnt_col, read) || read != cnt_row * cnt_col)
     {
       std::cerr << "ERROR!\n";
@@ -73,7 +73,7 @@ int main(int argc, char** argv)
       std::cerr << "Out of memory\n";
       return 1;
     }
-    int read = 0;
+    size_t read = 0;
     if (!shramko::inputMtrx(finput, mtrx, cnt_row, cnt_col, read) || read != cnt_row * cnt_col)
     {
       delete[] mtrx;
